refactor: rollback handler
